requirements.txt
datasets
pandas
PyYAML
soundfile
huggingface_hub